requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
PyYAML>=6.0.1
pyarrow>=12.0.0
tqdm>=4.66.0
rich>=13.0.0
clearml>=1.13.0
matplotlib>=3.7.0
seaborn>=0.13.0
datasets>=2.15.0